refactor to use doubles for prices

+++ src/main/java/behaviours/ISell.java
@@ -1,5 +1,5 @@
 package behaviours;
 
 public interface ISell {
-    long calculateMarkup();
+    double calculateMarkup();
 }
